+++ drivers/cxl/core/port.c
@@ -537,7 +537,7 @@ static void cxl_ep_remove(struct cxl_port *port, struct cxl_ep *ep)
 {
 	if (!ep)
 		return;
-	xa_erase(&port->endpoints, (unsigned long) ep->ep);
+	xa_erase(&port->endpoints, (uintptr_t) ep->ep);
 	cxl_ep_release(ep);
 }
 
@@ -983,13 +983,13 @@ struct pci_bus *cxl_port_to_pci_bus(struct cxl_port *port)
 		return pdev->subordinate;
 	}
 
-	return xa_load(&cxl_root_buses, (unsigned long)port->uport_dev);
+	return xa_load(&cxl_root_buses, (uintptr_t)port->uport_dev);
 }
 EXPORT_SYMBOL_NS_GPL(cxl_port_to_pci_bus, "CXL");
 
 static void unregister_pci_bus(void *uport_dev)
 {
-	xa_erase(&cxl_root_buses, (unsigned long)uport_dev);
+	xa_erase(&cxl_root_buses, (uintptr_t)uport_dev);
 }
 
 int devm_cxl_register_pci_bus(struct device *host, struct device *uport_dev,
@@ -1000,7 +1000,7 @@ int devm_cxl_register_pci_bus(struct device *host, struct device *uport_dev,
 	if (dev_is_pci(uport_dev))
 		return -EINVAL;
 
-	rc = xa_insert(&cxl_root_buses, (unsigned long)uport_dev, bus,
+	rc = xa_insert(&cxl_root_buses, (uintptr_t)uport_dev, bus,
 		       GFP_KERNEL);
 	if (rc)
 		return rc;
@@ -1067,7 +1067,7 @@ static int add_dport(struct cxl_port *port, struct cxl_dport *dport)
 		return -EBUSY;
 	}
 
-	rc = xa_insert(&port->dports, (unsigned long)dport->dport_dev, dport,
+	rc = xa_insert(&port->dports, (uintptr_t)dport->dport_dev, dport,
 		       GFP_KERNEL);
 	if (rc)
 		return rc;
@@ -1100,7 +1100,7 @@ static void cxl_dport_remove(void *data)
 	struct cxl_dport *dport = data;
 	struct cxl_port *port = dport->port;
 
-	xa_erase(&port->dports, (unsigned long) dport->dport_dev);
+	xa_erase(&port->dports, (uintptr_t) dport->dport_dev);
 	put_device(dport->dport_dev);
 }
 
@@ -1190,7 +1190,7 @@ __devm_cxl_add_dport(struct cxl_port *port, struct device *dport_dev,
 	    port->component_reg_phys != CXL_RESOURCE_NONE) {
 		rc = cxl_port_setup_regs(port, port->component_reg_phys);
 		if (rc) {
-			xa_erase(&port->dports, (unsigned long)dport->dport_dev);
+			xa_erase(&port->dports, (uintptr_t)dport->dport_dev);
 			return ERR_PTR(rc);
 		}
 		port->component_reg_phys = CXL_RESOURCE_NONE;
@@ -1290,7 +1290,7 @@ static int add_ep(struct cxl_ep *new)
 	if (port->dead)
 		return -ENXIO;
 
-	return xa_insert(&port->endpoints, (unsigned long)new->ep,
+	return xa_insert(&port->endpoints, (uintptr_t)new->ep,
 			 new, GFP_KERNEL);
 }
