+++ drivers/net/wireless/ath/ath12k/dp_tx.c
@@ -192,7 +192,7 @@ static int ath12k_dp_tx_align_payload(struct ath12k_base *ab,
 	int tailroom = skb_tailroom(skb);
 	int ret = 0;
 
-	offset = (unsigned long)skb->data & iova_mask;
+	offset = (uintptr_t)skb->data & iova_mask;
 	delta1 = offset;
 	delta2 = iova_mask - offset + 1;
 
@@ -209,7 +209,7 @@ static int ath12k_dp_tx_align_payload(struct ath12k_base *ab,
 
 		dev_kfree_skb_any(skb);
 
-		offset = (unsigned long)skb2->data & iova_mask;
+		offset = (uintptr_t)skb2->data & iova_mask;
 		if (offset)
 			ath12k_dp_tx_move_payload(skb2, offset, true);
 		*pskb = skb2;
