+++ include/linux/balloon_compaction.h
@@ -102,7 +102,7 @@ static inline void balloon_page_insert(struct balloon_dev_info *balloon,
 	__SetPageOffline(page);
 	if (IS_ENABLED(CONFIG_BALLOON_COMPACTION)) {
 		SetPageMovableOps(page);
-		set_page_private(page, (unsigned long)balloon);
+		set_page_private(page, (uintptr_t)balloon);
 	}
 	list_add(&page->lru, &balloon->pages);
 }
